chore(dev): pins the base-image version of the test Docker image (#74)
This change makes sure that we always use the same version of the `node` base-image for testing by making it explicit against which version of the image we're going to run our tests. 

Not specifying the version by explicit hash can for example cause issues when a local Docker cache has a very old version of the image, which causes the tests to fail.

In addition, we also configure Dependabot to automatically open version bump PRs for this testing image.