fix(timeout): resolve degredation which prevented setting indefinite timeout
The help text suggests that one should be able to configure an indefinite timeout by setting it to zero. However, the current implementation would immediately timeout while doing so.

This change ignores the timeout behaviour when the timeout argument is set to 0. In addition, a test is added which tests whether the timeout at least doesn't immediately time out.